Restructure Python code and run link check as cron job at Travis CI
